backend-devices/requirements.txt
blinker
certifi
charset-normalizer
click
Flask
Flask-SQLAlchemy
greenlet
idna
itsdangerous
Jinja2
MarkupSafe
pika
PyJWT
requests
SQLAlchemy
typing_extensions
urllib3
Werkzeug
psycopg2-binary
